requirements.txt
fastapi
uvicorn
httpx
google-auth
google-auth-oauthlib
google-api-python-client
openai
